+++ mains/common/header.h
@@ -3,6 +3,6 @@
 #define ROWS 10000
 #define COLS 5000
 #include <malloc.h>
-int *insert_arr(const int, const int);
-int *insert_from_file(char *, const int, const int);
+int *insert_arr(const size_t, const size_t);
+int *insert_from_file(char *, const size_t, const size_t);
 #endif
